Fix circular dependencies and first test
